Assignment2
This program takes a user entered value, check if it is an integer and performs a number of operations if it is an integer. If not a integer program asks user to read instructions carefully. The program prints the output of each operation on the screen and also checks if the magnitude of the number is greater than, equal to or less than 4. However, it does not tell user if the number entered is positive or negative.