Assignment 9
Ask user to enter file path.
Using Regex check to see if file path valid. Then check if File exists in directory. If file exist, determine word count of file. If it doesn't, using a do-while loop, ask user to enter valid path.  Program uses 3 methods and a main to accomplish this. The 3 methods are housed under the ProgramProcesses class. The main in located in the Program class.